EVM-1167: move MAX_EVENT_TOPICS to system constants
