+++ forward_system/src/run/result_keeper.rs
@@ -7,8 +7,8 @@ use zk_ee::common_structs::{
     GenericEventContent, GenericEventContentWithTxRef, GenericLogContent,
     GenericLogContentWithTxRef, PreimageType,
 };
-use zk_ee::storage_types::MAX_EVENT_TOPICS;
 use zk_ee::system::IOResultKeeper;
+use zk_ee::system::MAX_EVENT_TOPICS;
 use zk_ee::types_config::EthereumIOTypesConfig;
 use zk_ee::utils::{Bytes32, UsizeAlignedByteBox};
