+++ zk_ee/src/storage_types/mod.rs
@@ -1,9 +1,7 @@
 //! Serialization and deserialization helpers for storage-related keys and values.
 
-mod constants;
 mod initial_storage_slot_data;
 mod storage_address;
 
-pub use constants::MAX_EVENT_TOPICS;
 pub use initial_storage_slot_data::InitialStorageSlotData;
 pub use storage_address::StorageAddress;
